pyproject.toml
[project]
name = "aoai-func-cosmosdb"
version = "0.1.0"
description = ""
authors = [
    {name = "kimtth"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "azure-cosmos (>=4.9.0,<5.0.0)",
    "semantic-kernel (>=1.34.0,<2.0.0)",
    "azure-functions (>=1.23.0,<2.0.0)",
    "azure-identity (>=1.23.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
